Twitt/requirements.txt
djangorestframework
markdown
django-filter
djangorestframework-simplejwt==5.3.1